pyproject.toml
[project]
name = "nautilus-stoch-strategy"
version = "0.1.0"
description = "Stochastics trading strategy using NautilusTrader"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "nautilus_trader>=1.221.0",
    "pandas>=2.0.0",
    "yfinance>=0.2.0",
    "pyyaml>=6.0",
    "quantstats>=0.0.62",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
